pyproject.toml
[tool.poetry]
name = "pig"
version = "0.1.0"
description = ""
authors = ["Jim Cassidy <jim@ion8.net>"]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.6.1"
python-dotenv = "^0.21.0"
pip = "^22.2.2"
install = "^1.3.5"
google-cloud-storage = "^2.5.0"
git = "https://djamescassidy:Okrhcf90200!@github.com/djamescassidy/testcli.git"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
# Git dependencies



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
